+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [SW-690] Fixed a bug that caused products not to be exported that didn't have a manufacturer assigned.
 - [SW-691] Fixed a bug that caused internal additional listing requests to be blocked.
 - [SW-693] Fixed a bug that caused data of inactive and variants out-of-stock to be exported.
+- [SW-688] Fixed a bug that caused some routes to use MySQL instead of ElasticSearch for sales channels where Findologic was not active.
+- [SW-692] Fixed a bug that caused a no-result on category listings with Dynamic Product Groups.
 - [SW-687] Added Shopware 6.4.8.1 to the test matrix.
 
 # 2.6.1